refactor: split ChatPanel into smaller components
- Created shared types in types.ts
- Split MessageList into its own component
- Created ChatInput component
- Extracted logic into useChat hook
- Updated main ChatPanel to use new components
